--- a/build.gradle
+++ b/build.gradle
@@ -26,8 +26,7 @@ plugins {
   id 'io.spring.dependency-management' version '1.0.11.RELEASE'
   id 'org.springframework.boot' version '2.6.6'
   id 'com.github.ben-manes.versions' version '0.41.0'
-//  id 'hmcts.ccd.sdk' version '2.0.2'
-    id 'hmcts.ccd.sdk'
+  id 'hmcts.ccd.sdk' version '2.1.0'
   id 'com.github.hmcts.rse-cft-lib' version '0.15.62'
 }
 
@@ -121,6 +120,7 @@ task integration(type: Test) {
   failFast = true
 }
 
+// Invoked from CftLibConfig when booting with CCD.
 task buildCCDXlsx(type: Exec) {
   group 'ccd tasks'
   commandLine './bin/ccd-build-definition.sh'
@@ -316,6 +316,7 @@ dependencies {
   functionalTestImplementation group: 'com.github.hmcts', name: 'document-management-client', version: '7.0.1'
   functionalTestImplementation group: 'com.github.hmcts', name: 'service-auth-provider-java-client', version: '4.0.2'
 
+  // Provides fast-reload of just the NFDIV service.
   cftlibImplementation 'org.springframework.boot:spring-boot-devtools'
 }