--- a/settings.gradle
+++ b/settings.gradle
@@ -10,5 +10,3 @@ pluginManagement {
 }
 
 rootProject.name = 'nfdiv-case-api'
-
-includeBuild '../../ccd-config-generator'