--- a/src/cftlib/java/uk/gov/hmcts/divorce/cftlib/CftLibConfig.java
+++ b/src/cftlib/java/uk/gov/hmcts/divorce/cftlib/CftLibConfig.java
@@ -1,13 +1,16 @@
 package uk.gov.hmcts.divorce.cftlib;
 
 import org.apache.commons.io.IOUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.DefaultResourceLoader;
 import org.springframework.core.io.ResourceLoader;
 import org.springframework.stereotype.Component;
+import uk.gov.hmcts.ccd.sdk.ConfigWriter;
 import uk.gov.hmcts.rse.ccd.lib.api.CFTLib;
 import uk.gov.hmcts.rse.ccd.lib.api.CFTLibConfigurer;
 
+import java.io.File;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -18,6 +21,9 @@ public class CftLibConfig implements CFTLibConfigurer {
 
     @Value("ccd-NFD-${CCD_DEF_NAME:dev}.xlsx")
     String defName;
+
+    @Autowired
+    ConfigWriter configWriter;
 
     @Override
     public void configure(CFTLib lib) throws Exception {
@@ -52,6 +58,15 @@ public class CftLibConfig implements CFTLibConfigurer {
         var json = IOUtils.toString(resourceLoader.getResource("classpath:cftlib-am-role-assignments.json")
             .getInputStream(), Charset.defaultCharset());
         lib.configureRoleAssignments(json);
+
+        configWriter.generateAllCaseTypesToJSON(new File("build/definitions"));
+//        var code = Runtime.getRuntime().exec("./gradlew buildCCDXlsx").waitFor();
+        var code = new ProcessBuilder("./gradlew", "buildCCDXlsx")
+          .inheritIO()
+          .start().waitFor();
+      if (code != 0) {
+          throw new RuntimeException("Error converting ccd json to xlsx");
+        }
 
         var def = Files.readAllBytes(Path.of("build/ccd-config/" + defName));
         lib.importDefinition(def);