### Change description ###

Support fast-reloading of the NFDIV service during local development:

* Reload just the NFDIV service while leaving other CFT services running
* Regenerate ccd definitions on fast reboot (if necessary) inside CftlibConfig

Indicative performance benefits:

Scenario | master time | PR time |  % reduction
-- | -- | -- | --
Change a callback (no definition change) | 45 seconds | 6 seconds | 85%
Change ccd definition | 45 seconds | 24 seconds | 50%

Fast reload is provided by spring boot devtools; edit and compile one or more classes to trigger a fast restart.